Conflicts:
	composer.json
	config/module.config.php
	src/ZfcUserAdmin/Service/User.php
	view/zfc-user-admin/user-admin/list.phtml

+++ src/ZfcUserAdmin/Controller/UserAdminController.php
@@ -3,23 +3,30 @@
 namespace ZfcUserAdmin\Controller;
 
 use Zend\Mvc\Controller\AbstractActionController;
-use Zend\View\Model\ViewModel;
+use Zend\Paginator;
+use Zend\Stdlib\Hydrator\ClassMethods;
+use ZfcUser\Mapper\UserInterface;
+use ZfcUser\Options\ModuleOptions as ZfcUserModuleOptions;
 use ZfcUserAdmin\Options\ModuleOptions;
 
 class UserAdminController extends AbstractActionController
 {
-    protected $options, $userMapper, $adminUserService;
+    protected $options, $userMapper;
+    protected $zfcUserOptions;
+    /**
+     * @var \ZfcUserAdmin\Service\User
+     */
+    protected $adminUserService;
 
     public function listAction()
     {
         $userMapper = $this->getUserMapper();
         $users = $userMapper->findAll();
         if (is_array($users)) {
-            $paginator = new \Zend\Paginator\Paginator(new \Zend\Paginator\Adapter\ArrayAdapter($users));
+            $paginator = new Paginator\Paginator(new Paginator\Adapter\ArrayAdapter($users));
         } else {
             $paginator = $users;
         }
-
 
         $paginator->setItemCountPerPage(100);
         $paginator->setCurrentPageNumber($this->getEvent()->getRouteMatch()->getParam('p'));
@@ -31,55 +38,77 @@ class UserAdminController extends AbstractActionController
 
     public function createAction()
     {
+        /** @var $form \ZfcUserAdmin\Form\CreateUser */
         $form = $this->getServiceLocator()->get('zfcuseradmin_createuser_form');
         $request = $this->getRequest();
 
-        $user = false;
-        if($request->isPost())
-        {
-            $user = $this->getAdminUserService()->create((array) $request->getPost());
+        /** @var $request \Zend\Http\Request */
+        if ($request->isPost()) {
+            $zfcUserOptions = $this->getZfcUserOptions();
+            $class = $zfcUserOptions->getUserEntityClass();
+            $user = new $class();
+            $form->setHydrator(new ClassMethods());
+            $form->bind($user);
+            $form->setData($request->getPost());
+
+            if ($form->isValid()) {
+                $user = $this->getAdminUserService()->create($form, (array)$request->getPost());
+                if ($user) {
+                    $this->flashMessenger()->addSuccessMessage('The user was created');
+                    return $this->redirect()->toRoute('zfcadmin/zfcuseradmin/list');
+                }
+            }
         }
 
-        if (!$user) {
-            return array(
-                'createUserForm' => $form
-            );
-        }
-
-        $this->flashMessenger()->setNamespace('zfcuseradmin')->addMessage('The user was created');
-        return $this->redirect()->toRoute('zfcadmin/zfcuseradmin/list');
+        return array(
+            'createUserForm' => $form
+        );
     }
 
     public function editAction()
     {
         $userId = $this->getEvent()->getRouteMatch()->getParam('userId');
         $user = $this->getUserMapper()->findById($userId);
+
+        /** @var $form \ZfcUserAdmin\Form\EditUser */
         $form = $this->getServiceLocator()->get('zfcuseradmin_edituser_form');
         $form->setUser($user);
-        $request = $this->getRequest();
 
-        if (!$request->isPost()) {
+        /** @var $request \Zend\Http\Request */
+        $request = $this->getRequest();
+        if ($request->isPost()) {
+            $form->setData($request->getPost());
+            if ($form->isValid()) {
+                $user = $this->getAdminUserService()->edit($form, (array)$request->getPost(), $user);
+                if ($user) {
+                    $this->flashMessenger()->addSuccessMessage('The user was edited');
+                    return $this->redirect()->toRoute('zfcadmin/zfcuseradmin/list');
+                }
+            }
+        } else {
             $form->populateFromUser($user);
-            return array(
-                'editUserForm' => $form,
-                'userId' => $userId
-            );
         }
 
-        $this->getAdminUserService()->edit(get_object_vars($request->getPost()), $user);
-
-        $this->flashMessenger()->setNamespace('zfcuseradmin')->addMessage('The user was edited');
-        return $this->redirect()->toRoute('zfcadmin/zfcuseradmin/list');
+        return array(
+            'editUserForm' => $form,
+            'userId' => $userId
+        );
     }
 
     public function removeAction()
     {
         $userId = $this->getEvent()->getRouteMatch()->getParam('userId');
-        $user = $this->getUserMapper()->findById($userId);
-        if($user)
-        {
-            $this->getUserMapper()->remove($user);
-            $this->flashMessenger()->setNamespace('zfcuseradmin')->addMessage('The user was deleted');
+
+        /** @var $identity \ZfcUser\Entity\UserInterface */
+        $identity = $this->zfcUserAuthentication()->getIdentity();
+        if ($identity && $identity->getId() == $userId) {
+            $this->flashMessenger()->addErrorMessage('You can not delete yourself');
+        } else {
+            $user = $this->getUserMapper()->findById($userId);
+            if ($user) {
+                $this->getUserMapper()->remove($user);
+                $this->flashMessenger()->addSuccessMessage('The user was deleted');
+            }
         }
 
         return $this->redirect()->toRoute('zfcadmin/zfcuseradmin/list');
@@ -107,7 +136,7 @@ class UserAdminController extends AbstractActionController
         return $this->userMapper;
     }
 
-    public function setUserMapper(UserMapperInterface $userMapper)
+    public function setUserMapper(UserInterface $userMapper)
     {
         $this->userMapper = $userMapper;
         return $this;
@@ -125,5 +154,22 @@ class UserAdminController extends AbstractActionController
     {
         $this->adminUserService = $service;
         return $this;
+    }
+
+    public function setZfcUserOptions(ZfcUserModuleOptions $options)
+    {
+        $this->zfcUserOptions = $options;
+        return $this;
+    }
+
+    /**
+     * @return \ZfcUser\Options\ModuleOptions
+     */
+    public function getZfcUserOptions()
+    {
+        if (!$this->zfcUserOptions instanceof ZfcUserModuleOptions) {
+            $this->setZfcUserOptions($this->getServiceLocator()->get('zfcuser_module_options'));
+        }
+        return $this->zfcUserOptions;
     }
 }
